Update 0.1.0 - 010.711

+++ app/src/background/server/commands.js
@@ -31,8 +31,8 @@ module.exports = {
             if(para.includes("--alwaysstart")) serv.writeConfig("shouldRun", true)
             return serverShell.runSHELL(startLine)
          }
-         return false
       }
+      return false
    },
 
    /**
@@ -55,11 +55,55 @@ module.exports = {
                return serverShell.runSHELL(`kill ${info.pid}`)
             }
             else {
+               setTimeout(() => {
+                  if(serv.isrun()) serverShell.runSHELL(`kill ${info.pid}`)
+               }, 30000)
                return CommandUtil.sendToScreen(server, "stop")
             }
          }
-         return false
       }
+      return false
+   },
+
+   /**
+    * Startet einen Server neu (auch wenn dieser Offline ist!)
+    * @param server {string}
+    * @param para {array} Parameters
+    * <br>
+    * - **--hardstop** (Beendet mit kill) <br>
+    * - **--alwaysstart** (Server startet immer wenn dieser NICHT läuft) <br>
+    * @return {boolean}
+    */
+   doRestart: (server, para) => {
+      let serv       = new serverClass(server)
+      if(serv.serverExsists()) {
+         let interval         = undefined
+         let doStartInterval  = undefined
+
+         if(serv.isrun()) {
+            module.exports.doStop(server, para)
+            interval    = setInterval(() => {
+               if(!serv.isrun()) {
+                  clearInterval(interval)
+                  interval = undefined
+               }
+            }, 2000)
+         }
+
+         doStartInterval = setInterval(() => {
+            if(!interval) {
+               if(!serv.isrun()) {
+                  clearInterval(doStartInterval)
+                  doStartInterval = undefined
+
+                  module.exports.doStart(server, para)
+               }
+            }
+         }, 2000)
+
+         return true
+      }
+      return false
    },
 
    /**
@@ -109,8 +153,49 @@ module.exports = {
          ) return false
 
          if(globalUtil.safeFileCreateSync([backuprun]) && paths.length !== 0) {
-            serverShell.runSHELL(`cd ${servCFG.path} && zip -9 -r ${zipPath} ${paths.join(" ")} && rm ${backuprun}`)
-            return true
+            // prüfe backupverzeichnis
+            let checkBackupPath = function () {
+            let haveRm          = false
+
+            let maxSize   = servCFG.autoBackupMaxDirSize
+            let maxCount  = servCFG.autoBackupMaxCount
+            if(maxCount !== 0 || maxSize !== 0) {
+               maxSize              = maxSize * 1e+6
+               let backupDirInfos   = globalUtil.safeFileReadDirSync([servCFG.pathBackup])
+               let totalSize        = 0
+               let totalCount       = 0
+               let oldestFile       = false
+               for(let file of backupDirInfos) {
+                  if(file.FileExt === ".zip") {
+                     let backupTime = parseInt(file.namePure, 10)
+                     // finde Ältestes Backup
+                     if (backupTime < oldestFile || oldestFile === false)
+                        oldestFile = backupTime
+                     // zähle Backups
+                     totalCount++
+                     // erfasse TotalSize
+                     totalSize += file.sizebit
+                  }
+               }
+
+               if(oldestFile !== false)
+                  if(
+                     (maxCount !== 0 && maxCount <= totalCount) ||
+                     (maxSize !== 0 && maxSize <= totalSize)
+                  ) {
+                     globalUtil.safeFileRmSync([servCFG.pathBackup, `${oldestFile}.zip`])
+                     haveRm = true
+                  }
+               }
+
+               if(haveRm) checkBackupPath()
+               return true
+            }
+
+            if(checkBackupPath()) {
+               serverShell.runSHELL(`cd ${servCFG.path} && zip -9 -r ${zipPath} ${paths.join(" ")} && rm ${backuprun}`)
+               return true
+            }
          }
 
          return false
